backend/requirements.txt
librosa 
matplotlib 
IPython
soundfile
pydub
django==4.2.0
djangorestframework==3.14.0